requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9  # PostgreSQL
mysqlclient==2.2.0      # MySQL (alternative)

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Document Generation
WeasyPrint==60.1
Jinja2==3.1.2
Pillow==10.1.0
reportlab==4.0.7

# Payment Gateway
razorpay==1.4.1

# WhatsApp Integration
requests==2.31.0
python-decouple==3.8

# Email
django-ses==3.5.2

# File Handling
python-magic==0.4.27

# Task Queue (for async tasks)
celery==5.3.4
redis==5.0.1

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Charts & Analytics
pandas==2.1.3
numpy==1.26.2

# Environment Variables
python-dotenv==1.0.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
faker==20.1.0

# API Documentation
drf-yasg==1.21.7
